static extension : TensorFlowTransform
